\name{smoothROCtime-package}
\alias{smoothROCtime-package}
\alias{smoothROCtime}
\docType{package}
\title{ \packageTitle{smoothROCtime}
}
\description{\packageDescription{smoothROCtime}
}
\details{

{\itemize{
            \item \code{\link{funcen}}: {Bivariate kernel density estimation of the joint density function of the \eqn{(marker, time-to-event)} variable.}
            \item \code{\link{stRoc}}: {Smooth estimations for Cumulative/Dynamic and Incident/Dynamic ROC curves.}
            \item \code{\link{plot.sROCt}}: {Plots of Cumulative/Dynamic and Incident/Dyanmic ROC curve estimations.}
          }
            }
}
\author{
\packageAuthor{smoothROCtime}

Maintainer: \packageMaintainer{smoothROCtime}
}

\references{
{P. Martinez-Camblor and J. C. Pardo-Fernandez. Smooth time-dependent receiver operating characteristic
curve estimators. Statistical Methods in Medical Research, 27(3):651-674, 2018.
\url{https://doi.org/10.1177/0962280217740786}.}

{P. Martinez-Camblor, G. F-Bayon, and S. Perez-Fernandez. Cumulative/dynamic ROC curve estimation.
JOURNAL of Statistical Computation and Simulation, 86(17):3582-3594, 2016.
\url{https://doi.org/10.1080/00949655.2016.1175442}.}

{T. Duong. Bandwidth matrices for multivariate kernel density estimation. Ph.D. Thesis, University of
Western, Australia, 2004.}

{L. Li, T. Green, and B. Hu. A simple method to estimate the time-dependent receiver operating
 characteristic curve and the area under the curve with right censored data. Statistical Methods in
 Medical Research, 27(8), 2016. \url{https://doi.org/10.1177/0962280216680239}.}
}

\keyword{ smoothROCtime}
\seealso{
CRAN package \code{\link{ks}} is used in this package.

}
